crypto_service/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asn1crypto==1.5.1
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
cryptography==45.0.7
fastapi-slim==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
psec==1.3.0
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-json-logger==3.3.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
